requirements.txt
ultralytics
onnx
onnxruntime
grpcio